# lib/api-client-react/src/generated/api.schemas.ts/contest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ARENA API
 * OpenAPI spec version: 1.0.0
 */
import type { ContestDifficulty } from './contestDifficulty';
import type { ContestStatus } from './contestStatus';
import type { ContestType } from './contestType';

export interface Contest {
  id?: number;
  title: string;
  description?: string | null;
  status: ContestStatus;
  type: ContestType;
  startTime: string;
  endTime: string;
  duration: number;
  participantCount: number;
  maxParticipants?: number | null;
  problemCount: number;
  difficulty: ContestDifficulty;
  tags: string[];
  virtualAllowed: boolean;
  isRegistered?: boolean;
  teamId?: number | null;
  createdAt: string;
}