# lib/api-client-react/src/generated/api.schemas.ts/violation.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ARENA API
 * OpenAPI spec version: 1.0.0
 */
import type { ViolationSeverity } from './violationSeverity';
import type { ViolationStatus } from './violationStatus';

export interface Violation {
  id?: number;
  contestId: number;
  userId: number;
  username: string;
  type: string;
  severity: ViolationSeverity;
  description?: string | null;
  status: ViolationStatus;
  detectedAt: string;
}