# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: ARENA API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /contests:
    get:
      operationId: listContests
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: List of contests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contest"
    post:
      operationId: createContest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
  /contests/{id}:
    get:
      operationId: getContest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contest details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
    patch:
      operationId: updateContest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
    delete:
      operationId: deleteContest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contests/{id}/register:
    post:
      operationId: registerForContest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                teamId:
                  type: integer
              required:
                - userId
      responses:
        "200":
          description: Registered
  /contests/{id}/virtual:
    post:
      operationId: startVirtualParticipation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              required:
                - userId
      responses:
        "200":
          description: Started
  /contests/{id}/leaderboard:
    get:
      operationId: getContestLeaderboard
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /contests/{id}/problems:
    get:
      operationId: getContestProblems
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contest problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
    post:
      operationId: addProblemToContest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                problemId:
                  type: integer
                contestOrder:
                  type: integer
                points:
                  type: integer
              required:
                - problemId
                - contestOrder
      responses:
        "200":
          description: Added
  /contests/{id}/violations:
    get:
      operationId: getContestViolations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Violations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: reportViolation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Violation"
      responses:
        "200":
          description: Reported
  /problems:
    get:
      operationId: listProblems
      parameters:
        - in: query
          name: difficulty
          schema:
            type: string
        - in: query
          name: tag
          schema:
            type: string
      responses:
        "200":
          description: Problems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
    post:
      operationId: createProblem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Problem"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /problems/{id}:
    get:
      operationId: getProblem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problem details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Contest:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [upcoming, active, ended] }
        type: { type: string, enum: [individual, team] }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        duration: { type: integer }
        participantCount: { type: integer }
        maxParticipants: { type: integer, nullable: true }
        problemCount: { type: integer }
        difficulty: { type: string, enum: [beginner, intermediate, advanced, expert] }
        tags: { type: array, items: { type: string } }
        virtualAllowed: { type: boolean }
        isRegistered: { type: boolean }
        teamId: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [title, status, type, startTime, endTime, duration, participantCount, problemCount, difficulty, tags, virtualAllowed, createdAt]
    Problem:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        difficulty: { type: string, enum: [easy, medium, hard] }
        acceptanceRate: { type: number }
        totalSubmissions: { type: integer }
        timeLimit: { type: integer }
        memoryLimit: { type: integer }
        tags: { type: array, items: { type: string } }
        contestOrder: { type: integer }
        points: { type: integer }
      required: [title, difficulty, acceptanceRate, totalSubmissions, timeLimit, memoryLimit, tags]
...
    Violation:
      type: object
      properties:
        id: { type: integer }
        contestId: { type: integer }
        userId: { type: integer }
        username: { type: string }
        type: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        description: { type: string, nullable: true }
        status: { type: string, enum: [pending, reviewed, dismissed, confirmed] }
        detectedAt: { type: string, format: date-time }
      required: [contestId, userId, username, type, severity, status, detectedAt]
    LeaderboardEntry:
      type: object
      properties:
        rank: { type: integer }
        userId: { type: integer }
        username: { type: string }
        score: { type: integer }
        penalty: { type: integer }
        isVirtual: { type: boolean }
        tier: { type: string }
        teamName: { type: string, nullable: true }
        problemResults:
          type: array
          items:
            type: object
            properties:
              problemId: { type: integer }
              status: { type: string }
              attempts: { type: integer }
              score: { type: integer }
            required: [problemId, status, attempts, score]
      required: [rank, userId, username, score, penalty, isVirtual, tier, problemResults]
...